+++ lib/player.rb
@@ -1,5 +1,5 @@
 class Player
-  attr_accessor :name, :symbol
+  attr_reader :name, :symbol
 
   def initialize(name, symbol)
     @name = name
